+++ doomgeneric/i_timer.h
@@ -27,7 +27,7 @@
 int I_GetTime (void);
 
 // returns current time in ms
-int I_GetTimeMS (void);
+long long I_GetTimeMS (void);
 
 // Pause for a specified number of ms
 void I_Sleep(int ms);
